mess with syncing
config/data synced messages may be passed to multiple wrappers instead of only the first. useful if you end up creating separate wrappers  in client and server code, it will still cope with it in single player. might be a race condition if you change it again before the packet arrives but then you send another packet so i think it fixes itself